Integrating Google APIs

+++ PATHLY_API/Dto/TripDto.cs
@@ -7,7 +7,7 @@ namespace PATHLY_API.Dto
         public int Id { get; set; }
         public string StartLocation { get; set; }
         public string EndLocation { get; set; }
-        public decimal Distance { get; set; }
+        public double Distance { get; set; }
         public DateTime StartTime { get; set; } = DateTime.UtcNow;
         public DateTime? EndTime { get; set; }
         public TripStatus Status { get; set; }
